Update README.md
There is the option to jump right up to jruby-9.0.5.0-complete, and I guess I might do that before doing merge with master and doing a release. But if the jruby team decided to require jdk8 at some stage that might not be a such a good idea (but they will probably defer that to jruby-9.1+ by which time most people will be using JRubyArt instead, if not why not)?